Разделила код на отдельные ES-модули
